+++ Civi/DataProcessor/Output/UIOutputHelper.php
@@ -15,7 +15,7 @@ namespace Civi\DataProcessor\Output;
  */
 class UIOutputHelper {
 
-  private static $rebuildMenu = false;
+  private static $rebuildMenu = true;
 
   /**
    * Delegation of the alter menu hook. Add the search outputs to the menu system.
@@ -56,92 +56,59 @@ class UIOutputHelper {
     }
   }
 
-  /**
-   * Update the navigation data when an output is saved/deleted from the database.
-   *
-   * @param $op
-   * @param $objectName
-   * @param $objectId
-   * @param $objectRef
-   */
-  public static function preHook($op, $objectName, $id, &$params) {
-    if ($objectName == 'DataProcessorOutput') {
-      // Disable this hook in unit tests because the menu rebuild it causes breaks transactions.
-      if (CIVICRM_UF === 'UnitTests') {
-        return;
-      }
-      $factory = dataprocessor_get_factory();
-      if ($op == 'delete') {
-        $output = civicrm_api3('DataProcessorOutput', 'getsingle', ['id' => $id]);
-        $outputClass = $factory->getOutputByName($output['type']);
-        if ($outputClass instanceof UIOutputInterface) {
-          self::$rebuildMenu = TRUE;
-        }
-        self::removeOutputFromNavigation($output['configuration']);
-      }
-      elseif ($op == 'edit') {
-        $output = civicrm_api3('DataProcessorOutput', 'getsingle', ['id' => $id]);
-        $outputClass = $factory->getOutputByName($output['type']);
-        if ($outputClass instanceof UIOutputInterface) {
-          self::$rebuildMenu = TRUE;
-        }
-        if (!isset($output['configuration']['navigation_id']) && !isset($params['configuration']['navigation_parent_path'])) {
-          return;
-        }
-        elseif (!isset($params['configuration']['navigation_parent_path'])) {
-          self::removeOutputFromNavigation($output['configuration']);
-        }
-        elseif (isset($params['configuration']['navigation_parent_path'])) {
-          if (!isset($output['configuration']['navigation_parent_path']) || $params['configuration']['navigation_parent_path'] != $output['configuration']['navigation_parent_path']) {
-            // Merge the current output from the database with the updated values
-            $configuration = array_merge($output['configuration'], $params['configuration']);
-            $configuration['navigation_parent_path'] = $params['configuration']['navigation_parent_path'];
-            $output = array_merge($output, $params);
-            $output['configuration'] = $configuration;
-            $dataProcessor = civicrm_api3('DataProcessor', 'getsingle', ['id' => $output['data_processor_id']]);
-            $configuration = self::createOrUpdateNavigationItem($output, $dataProcessor);
-            if ($configuration) {
-              $params['configuration'] = $configuration;
-            }
+  public static function navigationMenuHook(&$menu) {
+    $factory = dataprocessor_get_factory();
+    $dao = \CRM_Core_DAO::executeQuery("
+      SELECT DISTINCT o.id as output_id, o.type, p.id as dataprocessor_id
+      FROM civicrm_data_processor_output o
+      INNER JOIN civicrm_data_processor p ON o.data_processor_id = p.id
+      WHERE p.is_active = 1 and p.id = 9
+    ");
+    while ($dao->fetch()) {
+      $outputClass = $factory->getOutputByName($dao->type);
+      if ($outputClass instanceof UIOutputInterface) {
+        $output = civicrm_api3('DataProcessorOutput', 'getsingle', ['id' => $dao->output_id]);
+        self::fixBackwardsCompatibility($output);
+        if (isset($output['configuration']['navigation_parent_path'])) {
+          $dataProcessor = civicrm_api3('DataProcessor', 'getsingle', ['id' => $dao->dataprocessor_id]);
+          $title = $dataProcessor['title'];
+          if ($outputClass && $outputClass instanceof \Civi\DataProcessor\Output\UIFormOutputInterface) {
+            $url = $outputClass->getUrlToUi($output, $dataProcessor);
+            $title = $outputClass->getTitleForUiLink($output, $dataProcessor);
           }
-        }
-      }
-      elseif ($op == 'create') {
-        self::$rebuildMenu = true;
-        if (isset($params['configuration']['navigation_parent_path'])) {
-          $dataProcessor = civicrm_api3('DataProcessor', 'getsingle', ['id' => $params['data_processor_id']]);
-          $configuration = self::createOrUpdateNavigationItem($params, $dataProcessor);
-          if ($configuration) {
-            $params['configuration'] = $configuration;
+          $item = [
+            'label' => $title,
+            'name' => $dataProcessor['name'].'_'.$output['type'],
+            'url' => \CRM_Utils_System::url($url, 'reset=1', true),
+            'operator' => 'OR',
+            'separator' => 0,
+          ];
+          if (isset($output['permission'])) {
+            $item['permission'] = $output['permission'];
           }
-        }
-      }
-    } elseif ($objectName == 'DataProcessor' && $op == 'edit') {
-      $dataProcessor = civicrm_api3('DataProcessor', 'getsingle', ['id' => $id]);
-      if (isset($params['is_active']) && $params['is_active'] != $dataProcessor['is_active']) {
-        // Only update navigation when is active is changed
-        $dataProcessor = array_merge($dataProcessor, $params);
-        $outputs = civicrm_api3('DataProcessorOutput', 'get', ['data_processor_id' => $id, 'options' => ['limit' => 0]]);
-        foreach($outputs['values'] as $output) {
-          if (isset($output['configuration']['navigation_id'])) {
-            self::createOrUpdateNavigationItem($output, $dataProcessor);
-          }
+          _dataprocessor_civix_insert_navigation_menu($menu, $output['configuration']['navigation_parent_path'], $item);
         }
       }
     }
   }
 
   /**
-   * Remove an output from the navigation menu.
+   * Fix backwards compatibility. Previously we stored the navigation id in the database.
+   * Now we have the parent path.
    *
-   * @param array $configuration
+   * @param $output
    */
-  private static function removeOutputFromNavigation($configuration) {
-    if (isset($configuration['navigation_id'])) {
-      $navId = $configuration['navigation_id'];
-      \CRM_Core_BAO_Navigation::processDelete($navId);
-      \CRM_Core_BAO_Navigation::resetNavigation();
-      self::$rebuildMenu = TRUE;
+  public static function fixBackwardsCompatibility($output) {
+    $navigation = \CRM_Dataprocessor_Utils_Navigation::singleton();
+    if (isset($output['configuration']['navigation_id'])) {
+      // Backwards compatibility.
+      $output['configuration']['navigation_parent_path'] = $navigation->getNavigationParentPathById($output['configuration']['navigation_id']);
+      \CRM_Core_BAO_Navigation::processDelete($output['configuration']['navigation_id']);
+      unset($output['configuration']['navigation_id']);
+      \CRM_Core_DAO::executeQuery("UPDATE civicrm_data_processor_output SET configuration = %1 WHERE id = %2", [
+        1 => [json_encode($output['configuration']), 'String'],
+        2 => [$output['id'],'Integer'],
+      ]);
     }
   }
 
@@ -154,21 +121,13 @@ class UIOutputHelper {
    * @param $objectRef
    */
   public static function postHook($op, $objectName, $id, &$objectRef) {
-    if ($objectName != 'DataProcessorOutput') {
-      return;
-    }
     // Disable this hook in unit tests because the menu rebuild it causes breaks transactions.
     if (CIVICRM_UF === 'UnitTests') {
       return;
     }
 
-    if (self::$rebuildMenu) {
-      // Rebuild the CiviCRM Menu (which holds all the pages)
-      \CRM_Core_Menu::store();
-
-      // also reset navigation
-      \CRM_Core_BAO_Navigation::resetNavigation();
-      self::$rebuildMenu = false;
+    if (self::$rebuildMenu && ($objectName == 'DataProcessorOutput' || $objectName == 'DataProcessor')) {
+      self::rebuildMenuAndNavigation();
     }
   }
 
@@ -187,67 +146,21 @@ class UIOutputHelper {
   }
 
   /**
-   * Inserts/updates an navigation item.
-   * @param $params
-   * @param $dataProcessor
-   *
-   * @return array
+   * Disable the rebuilding of the menu and navigation.
+   * Usefull during an batch import.
    */
-  private static function createOrUpdateNavigationItem($output, $dataProcessor) {
-    $url = "";
-    $factory = dataprocessor_get_factory();
-    if (!isset($output['type'])) {
-      return false;
-    }
-    $outputClass = $factory->getOutputByName($output['type']);
-    $configuration = $output['configuration'];
+  public static function disableRebuildingOfMenuAndNavigation() {
+    self::$rebuildMenu = false;
+  }
 
-    $title = $dataProcessor['title'];
-    if ($outputClass && $outputClass instanceof \Civi\DataProcessor\Output\UIFormOutputInterface) {
-      $url = $outputClass->getUrlToUi($output, $dataProcessor);
-      $title = $outputClass->getTitleForUiLink($output, $dataProcessor);
-    }
-
-    $navigation = \CRM_Dataprocessor_Utils_Navigation::singleton();
-    $navigationParams = array();
-    // Retrieve the current navigation ID.
-    if (isset($configuration['navigation_id'])) {
-      // Get the default navigation parent id.
-      $navigationDefaults = [];
-      $retrieveNavParams = ['id' => $configuration['navigation_id']];
-      \CRM_Core_BAO_Navigation::retrieve($retrieveNavParams, $navigationDefaults);
-      if (!empty($navigationDefaults['id'])) {
-        $navigationParams['id'] = $navigationDefaults['id'];
-        $navigationParams['current_parent_id'] = !empty($navigationDefaults['parent_id']) ? $navigationDefaults['parent_id'] : NULL;
-        $navigationParams['parent_id'] = !empty($navigationDefaults['parent_id']) ? $navigationDefaults['parent_id'] : NULL;
-      }
-    }
-
-    $navigationParams['domain_id'] = \CRM_Core_Config::domainID();
-    $navigationParams['permission'] = array();
-    $navigationParams['label'] = $title;
-    $navigationParams['name'] = $dataProcessor['name'];
-
-    if (isset($configuration['navigation_parent_path'])) {
-      $navigationParams['parent_id'] = $navigation->getNavigationIdByPath($configuration['navigation_parent_path']);
-    }
-    $navigationParams['is_active'] = $dataProcessor['is_active'];
-
-    if (isset($output['permission'])) {
-      $navigationParams['permission'][] = $output['permission'];
-    }
-
-    unset($configuration['navigation_parent_path']);
-
-    $navigationParams['url'] = $url.'?reset=1';
-    $navigation = \CRM_Core_BAO_Navigation::add($navigationParams);
+  /**
+   * Rebuild the menu and navigation.
+   */
+  public static function rebuildMenuAndNavigation() {
+    // Rebuild the CiviCRM Menu (which has the url and routing information of the pages).
+    \CRM_Core_Menu::store();
+    // Also reset navigation
     \CRM_Core_BAO_Navigation::resetNavigation();
-
-    $configuration['navigation_id'] = $navigation->id;
-
-    self::$rebuildMenu = TRUE;
-
-    return $configuration;
   }
 
 }